JMX communication changes for the sidecar mode
I added a set of accessible datacenters that are useful to determine if a node is accessible from a Reaper instance in EACH mode notably. The set is populated during the JMX connection initialization phase where we try to connect to all nodes.
The active compactions are now polled and stored into the backend every 10 seconds so they can be retrieved from any reaper in sidecar mode.